Add files via upload
Added feature where user inputs streets and neighborhoods that he wants the script to monitor.
Updated parsing logic.
Updated Email notifications content. Now the mail won't be sent to user if the streets that he selected aren't mentioned in notifications in official site. Municipalities and streets that user selected will be highlighted in the mail.
Added Scheduler that will run checks on every 24 hours in background, instead of time.sleep()